# admin-api/src/main/resources/api/openapi.yaml
openapi: 3.0.3
info:
  title: Unchain Admin API
  description: |
    Comprehensive OpenAPI specification for the Unchain Admin API.

    This API controls the configuration of the Unchain Feature Management platform, 
    including Features, Strategies, Projects, Environments, Context Fields, 
    Change Requests, Segments, FeatureTags, and Dependencies.

    **Base URL**: `<your-unchain-instance>/api/admin`
  version: 4.0.0
  contact:
    name: Philipp Walther
servers:
  - url: http://localhost:8080
    description: Local Instance
security:
  - OidcAuth: []
...
paths:
...
  /context:
    get:
      operationId: listContextFields
      summary: Get all context fields
      tags: [Context]
      responses:
        "200":
          description: List of context fields
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContextField"
    post:
      operationId: createContextField
      summary: Create a context field
      tags: [Context]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContextField"
      responses:
        "201":
          description: Context field created
  /context/{contextField}:
    put:
      operationId: updateContextField
      summary: Update a context field
      tags: [Context]
      parameters:
        - name: contextField
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContextField"
      responses:
        "200":
          description: Context field updated
    delete:
      operationId: deleteContextField
      summary: Delete a context field
      tags: [Context]
      parameters:
        - name: contextField
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Context field deleted
  /context/{contextField}/legal-values:
    post:
      operationId: addLegalValueToContextField
      summary: Add a legal value to a context field
      tags: [Context]
      parameters:
        - name: contextField
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LegalValue"
      responses:
        "200":
          description: Legal value added
  /context/{contextField}/legal-values/{value}:
    delete:
      operationId: removeLegalValueFromContextField
      summary: Remove a legal value from a context field
      tags: [Context]
      parameters:
        - name: contextField
          in: path
          required: true
          schema:
            type: string
        - name: value
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Legal value removed

  # =========================================================================
  # GLOBAL STRATEGY DEFINITIONS
  # =========================================================================
...
components:
  securitySchemes:
    OidcAuth:
      type: openIdConnect
      openIdConnectUrl: https://server.redmoon.ch/auth/oidc/.well-known/openid-configuration
...
  schemas:
...
    ContextField:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        description:
          type: string
        stickiness:
          type: boolean
        sortOrder:
          type: integer
        legalValues:
          type: array
          items:
            $ref: "#/components/schemas/LegalValue"
    LegalValue:
      type: object
      properties:
        value:
          description: a value that is allowed for this context field
          type: string
          example: "fish"
        description:
          description: description of the legal value
          type: string
          example: "lives in the sea"

    # --- Change Request Schemas ---
...